backend/requirements.txt
fastapi
uvicorn
pydantic
scikit-learn
pandas
numpy
python-dotenv
httpx
